requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.6.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.1.0
